refactor(store): instrument the miss instead of re-driving it
Retry exhaustion has never been observed - both confirmed cases were the
channel-open race, which the 500ms retry already rescues. Convergence was built
for a case inferred from code, at the cost of a widened connection contract and
two bugs the review caught in it.

Replaces it with the measurement that was missing: subscribeRequestRetry when an
attempt goes unanswered, subscribeRequestUnanswered when none are. Keeps the two
fixes that stand on their own - an error reply is no longer recorded as applied
state, and setAudio no longer leaves the desired state behind when it dedupes.
